+++ decidim-initiatives/app/models/decidim/initiatives_type.rb
@@ -4,6 +4,7 @@ module Decidim
   # Initiative type.
   class InitiativesType < ApplicationRecord
     validates :title, :description, presence: true
+    validates :online_signature_enabled, inclusion: { in: [true, false] }
 
     mount_uploader :banner_image, Decidim::BannerImageUploader
 
@@ -16,5 +17,20 @@ module Decidim
              class_name: "Decidim::InitiativesTypeScope",
              dependent: :destroy,
              inverse_of: :type
+
+    def allowed_signature_types_for_initiatives
+      signature_types = []
+
+      signature_types << "online" if Decidim::Initiatives.online_voting_allowed && online_signature_enabled
+      signature_types << "offline" if Decidim::Initiatives.face_to_face_voting_allowed
+      signature_types << "any" if signature_types.size == (Initiative.signature_types.size - 1)
+
+      signature_types
+    end
+
+    def initiatives
+      initiatives_ids = scopes.map { |scope| scope.initiatives.pluck(:id) }.flatten
+      Initiative.where(id: initiatives_ids)
+    end
   end
 end